deprecate: remove wildloop v1 code
Backend:
- Delete server/wild_loop.py (2118 lines) and tests/wild_loop_test.py (781 lines)
- Delete 7 v1 prompt skills (wild_alert, wild_analyzing, wild_exploring, etc.)
- Remove v1 imports, engine setup, and ~180 lines of v1 routes from server.py
- Inline WildModeRequest and wild_mode_enabled into server.py
- Remove auto_enqueue_alert, auto_enqueue_run_terminal, record_created_entity calls
- Remove _WILD_SENTINEL and build_wild_prompt from _build_chat_prompt

Frontend:
- Remove ~460 lines of v1 API functions/types from lib/api.ts
- Remove v1 enqueueWildEvent usage from hooks/use-wild-loop.ts
- Remove v1 imports, state, polling, ~280 lines of v1 JSX from wild-loop-debug-panel.tsx
- Remove dead WildLoopState interface from lib/types.ts
- Remove dead phaseColor, priorityLabel helpers

Verified: 39/39 v2 tests pass, npm build succeeds
